--- internal/group/embed.go
@@ -1,6 +1,0 @@
-package group
-
-import _ "embed"
-
-//go:embed sql/schema.sql
-var GroupDDL string
