+++ internal/group/sql/embed.go
@@ -1,0 +1,6 @@
+package group
+
+import _ "embed"
+
+//go:embed schema.sql
+var GroupSchema string
