+++ internal/metadata/sql/embed.go
@@ -1,0 +1,6 @@
+package metadata
+
+import _ "embed"
+
+//go:embed schema.sql
+var MetadataSchema string
